pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "structured-ld-agentic-rag"
version = "0.1.0"
description = "Structured Linked Data for Agentic RAG — ISWC 2026 research project"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Andrea Volpini" },
]

dependencies = [
    # Google Cloud / Vertex AI
    "google-cloud-aiplatform>=1.74.0",
    "google-cloud-vectorsearch>=0.1.0",
    "google-adk>=0.5.0",
    "google-genai>=1.5.0",

    # Data processing
    "beautifulsoup4>=4.12.0",
    "rdflib>=7.0.0",
    "lxml>=5.0.0",
    "httpx>=0.27.0",

    # Templating
    "jinja2>=3.1.0",

    # Analysis
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",

    # Configuration
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
